URSA-7914 | Addressing PRs comments
