+++ lib/src/rtc_session.dart
@@ -15,6 +15,7 @@ import 'name_addr_header.dart';
 import 'request_sender.dart';
 import 'rtc_session/dtmf.dart' as RTCSession_DTMF;
 import 'rtc_session/dtmf.dart';
+import 'rtc_session/ice_summary.dart';
 import 'rtc_session/info.dart' as RTCSession_Info;
 import 'rtc_session/info.dart';
 import 'rtc_session/refer_notifier.dart';
@@ -48,202 +49,6 @@ const Duration kIceRestartRetryWindow = Duration(seconds: 180);
 const Duration kIceRestartDebounce = Duration(seconds: 3);
 
 const Duration kIceCandidateSettleDelay = Duration(milliseconds: 250);
-
-class _IceServerSummary {
-  _IceServerSummary({
-    required this.count,
-    required this.stunConfigured,
-    required this.turnConfigured,
-  });
-
-  factory _IceServerSummary.fromPcConfig(Map<String, dynamic>? pcConfig) {
-    dynamic configuredServers = pcConfig?['iceServers'];
-    List<dynamic> servers = configuredServers is Iterable
-        ? configuredServers.toList()
-        : configuredServers == null
-            ? <dynamic>[]
-            : <dynamic>[configuredServers];
-    bool stunConfigured = false;
-    bool turnConfigured = false;
-
-    for (dynamic server in servers) {
-      dynamic configuredUrls = server is Map
-          ? server['urls'] ?? server['url']
-          : server;
-      List<dynamic> urls = configuredUrls is Iterable && configuredUrls is! String
-          ? configuredUrls.toList()
-          : configuredUrls == null
-              ? <dynamic>[]
-              : <dynamic>[configuredUrls];
-      for (dynamic value in urls) {
-        String url = value.toString().toLowerCase();
-        stunConfigured =
-            stunConfigured || url.startsWith('stun:') || url.startsWith('stuns:');
-        turnConfigured =
-            turnConfigured || url.startsWith('turn:') || url.startsWith('turns:');
-      }
-    }
-
-    return _IceServerSummary(
-      count: servers.length,
-      stunConfigured: stunConfigured,
-      turnConfigured: turnConfigured,
-    );
-  }
-
-  final int count;
-  final bool stunConfigured;
-  final bool turnConfigured;
-
-  String get logFields =>
-      'iceServerCount=$count stunConfigured=$stunConfigured '
-      'turnConfigured=$turnConfigured';
-}
-
-class _IceSdpSummary {
-  _IceSdpSummary({
-    required this.total,
-    required this.host,
-    required this.srflx,
-    required this.relay,
-    required this.prflx,
-    required this.unknown,
-    required this.udp,
-    required this.tcp,
-    required this.unknownProtocol,
-    required this.candidatesByMid,
-  });
-
-  factory _IceSdpSummary.fromSdp(String? sdp) {
-    int total = 0;
-    int host = 0;
-    int srflx = 0;
-    int relay = 0;
-    int prflx = 0;
-    int unknown = 0;
-    int udp = 0;
-    int tcp = 0;
-    int unknownProtocol = 0;
-    int currentMediaSection = -1;
-    int sessionLevelCandidates = 0;
-    List<int> candidateCountsBySection = <int>[];
-    List<String?> midsBySection = <String?>[];
-    Map<String, int> candidatesByMid = <String, int>{};
-
-    for (String rawLine in (sdp ?? '').split(RegExp(r'\r?\n'))) {
-      String line = rawLine.trim();
-      if (line.startsWith('m=')) {
-        currentMediaSection += 1;
-        candidateCountsBySection.add(0);
-        midsBySection.add(null);
-        continue;
-      }
-      if (line.startsWith('a=mid:')) {
-        if (currentMediaSection >= 0) {
-          midsBySection[currentMediaSection] =
-              line.substring('a=mid:'.length);
-        }
-        continue;
-      }
-      if (!line.startsWith('a=candidate:')) {
-        continue;
-      }
-
-      total += 1;
-      if (currentMediaSection >= 0) {
-        candidateCountsBySection[currentMediaSection] += 1;
-      } else {
-        sessionLevelCandidates += 1;
-      }
-      switch (candidateType(line)) {
-        case 'host':
-          host += 1;
-          break;
-        case 'srflx':
-          srflx += 1;
-          break;
-        case 'relay':
-          relay += 1;
-          break;
-        case 'prflx':
-          prflx += 1;
-          break;
-        default:
-          unknown += 1;
-      }
-      switch (candidateProtocol(line)) {
-        case 'udp':
-          udp += 1;
-          break;
-        case 'tcp':
-          tcp += 1;
-          break;
-        default:
-          unknownProtocol += 1;
-      }
-    }
-
-    if (sessionLevelCandidates > 0) {
-      candidatesByMid['none'] = sessionLevelCandidates;
-    }
-    for (int index = 0; index < candidateCountsBySection.length; index += 1) {
-      int count = candidateCountsBySection[index];
-      if (count == 0) {
-        continue;
-      }
-      String? mid = midsBySection[index];
-      String key = mid == null || mid.isEmpty ? 'm$index' : mid;
-      candidatesByMid[key] = (candidatesByMid[key] ?? 0) + count;
-    }
-
-    return _IceSdpSummary(
-      total: total,
-      host: host,
-      srflx: srflx,
-      relay: relay,
-      prflx: prflx,
-      unknown: unknown,
-      udp: udp,
-      tcp: tcp,
-      unknownProtocol: unknownProtocol,
-      candidatesByMid: candidatesByMid,
-    );
-  }
-
-  static final RegExp _candidateTypePattern =
-      RegExp(r'(?:^|\s)typ\s+(host|srflx|relay|prflx)(?:\s|$)');
-  static final RegExp _candidateProtocolPattern =
-      RegExp(r'^(?:a=)?candidate:\S+\s+\S+\s+(udp|tcp)(?:\s|$)',
-          caseSensitive: false);
-
-  static String candidateType(String? candidate) {
-    RegExpMatch? match =
-        _candidateTypePattern.firstMatch(candidate ?? '');
-    return match?.group(1) ?? 'unknown';
-  }
-
-  static String candidateProtocol(String? candidate) {
-    RegExpMatch? match =
-        _candidateProtocolPattern.firstMatch(candidate ?? '');
-    return match?.group(1)?.toLowerCase() ?? 'unknown';
-  }
-
-  final int total;
-  final int host;
-  final int srflx;
-  final int relay;
-  final int prflx;
-  final int unknown;
-  final int udp;
-  final int tcp;
-  final int unknownProtocol;
-  final Map<String, int> candidatesByMid;
-
-  String get logFields =>
-      'total=$total host=$host srflx=$srflx relay=$relay '
-      'prflx=$prflx unknown=$unknown udp=$udp tcp=$tcp '
-      'unknownProtocol=$unknownProtocol mids=$candidatesByMid';
-}
 
 class SIPTimers {
   Timer? ackTimer;
@@ -2026,8 +1831,8 @@ class RTCSession extends EventManager implements Owner {
     _pcConfig = Map<String, dynamic>.from(pcConfig);
     _rtcConstraints = Map<String, dynamic>.from(rtcConstraints);
 
-    _IceServerSummary iceServerSummary =
-        _IceServerSummary.fromPcConfig(pcConfig);
+    IceServerSummary iceServerSummary =
+        IceServerSummary.fromPcConfig(pcConfig);
     logger.i('createPeerConnection | creating RTCPeerConnection');
     logger.i(
         'peer_connection_constraints | ${iceServerSummary.logFields} '
@@ -2072,6 +1877,14 @@ class RTCSession extends EventManager implements Owner {
     logger.d('emit "peerconnection"');
     emit(EventPeerConnection(_connection));
     return;
+  }
+
+  String _callIdForLogging() {
+    try {
+      return _request?.call_id?.toString() ?? 'unknown';
+    } catch (_) {
+      return 'unknown';
+    }
   }
 
   Future<RTCSessionDescription> _createLocalDescription(
@@ -2129,16 +1942,8 @@ class RTCSession extends EventManager implements Owner {
     RTCPeerConnection connection = _connection!;
     Timer? iceCandidateSettleTimer;
     Future<void> readinessQueue = Future<void>.value();
-    _IceServerSummary iceServerSummary =
-        _IceServerSummary.fromPcConfig(_pcConfig);
-
-    String callId() {
-      try {
-        return _request?.call_id?.toString() ?? 'unknown';
-      } catch (_) {
-        return 'unknown';
-      }
-    }
+    IceServerSummary iceServerSummary =
+        IceServerSummary.fromPcConfig(_pcConfig);
 
     void cleanupIceGathering() {
       iceGatheringTimer?.cancel();
@@ -2167,7 +1972,7 @@ class RTCSession extends EventManager implements Owner {
         localDescription = await connection.getLocalDescription();
       } catch (error) {
         logger.w(
-            'local_description | decision=deferred callId=${callId()} '
+            'local_description | decision=deferred callId=${_callIdForLogging()} '
             'type=$type trigger=$trigger reason=description_read_failed '
             'elapsedMs=${iceGatheringStopwatch.elapsedMilliseconds} '
             'errorType=${error.runtimeType}');
@@ -2177,14 +1982,14 @@ class RTCSession extends EventManager implements Owner {
       }
       if (localDescription == null) {
         logger.w(
-            'local_description | decision=deferred callId=${callId()} '
+            'local_description | decision=deferred callId=${_callIdForLogging()} '
             'type=$type trigger=$trigger reason=missing_description '
             'elapsedMs=${iceGatheringStopwatch.elapsedMilliseconds}');
         return;
       }
 
-      _IceSdpSummary summary =
-          _IceSdpSummary.fromSdp(localDescription.sdp);
+      IceSdpSummary summary =
+          IceSdpSummary.fromSdp(localDescription.sdp);
       String? deferredReason;
       if (!force && summary.total == 0) {
         deferredReason = 'no_candidates';
@@ -2198,7 +2003,7 @@ class RTCSession extends EventManager implements Owner {
 
       if (deferredReason != null) {
         logger.i(
-            'local_description | decision=deferred callId=${callId()} '
+            'local_description | decision=deferred callId=${_callIdForLogging()} '
             'type=$type trigger=$trigger reason=$deferredReason '
             'elapsedMs=${iceGatheringStopwatch.elapsedMilliseconds} '
             'turnConfigured=${iceServerSummary.turnConfigured} '
@@ -2211,7 +2016,7 @@ class RTCSession extends EventManager implements Owner {
       _iceGatheringState = RTCIceGatheringState.RTCIceGatheringStateComplete;
       _rtcReady = true;
       logger.i(
-          'local_description | decision=ready callId=${callId()} type=$type '
+          'local_description | decision=ready callId=${_callIdForLogging()} type=$type '
           'trigger=$trigger elapsedMs=${iceGatheringStopwatch.elapsedMilliseconds} '
           'turnConfigured=${iceServerSummary.turnConfigured} '
           'callbacks=$candidateCallbacksBeforeReady ${summary.logFields}');
@@ -2280,11 +2085,11 @@ class RTCSession extends EventManager implements Owner {
         if (candidate != null) {
           candidateCallbacksBeforeReady += 1;
           logger.i(
-              'ice_candidate | callId=${callId()} '
+              'ice_candidate | callId=${_callIdForLogging()} '
               'sequence=$candidateCallbacksBeforeReady '
               'elapsedMs=${iceGatheringStopwatch.elapsedMilliseconds} '
-              'type=${_IceSdpSummary.candidateType(candidate.candidate)} '
-              'protocol=${_IceSdpSummary.candidateProtocol(candidate.candidate)} '
+              'type=${IceSdpSummary.candidateType(candidate.candidate)} '
+              'protocol=${IceSdpSummary.candidateProtocol(candidate.candidate)} '
               'sdpMid=${candidate.sdpMid} '
               'sdpMLineIndex=${candidate.sdpMLineIndex}');
           emit(EventIceCandidate(candidate, requestReadyAfterSettle));
